Chat_bot_ddasom_v1.1.0/Chat_bot_ddasom_v1.1.0/requirements.txt
streamlit
openai
python-dotenv
elevenlabs
langchain-community
langchain_openai
supabase
wordcloud
matplotlib
pandas
st_tailwind